feat(random-list.php): add redirect option for image requests
Add a new redirect feature to the random-list.php API. Clients can now specify
a 'redirect' parameter in their GET requests to control how image links are
handled. By default, the API performs a 302 redirect to the image link,
but if 'redirect' is set to 'stream', the image is sent using streamRemoteImage
functionality. This enhancement allows for more flexible image delivery
options to clients.